# tests/Forge.ReferenceCatalog.Tests/openapi.v1.snapshot.json
{
  "components": {
    "schemas": {
      "CatalogItemResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "message"
        ],
        "type": "object"
      },
      "CreateCatalogItemRequest": {
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "HttpValidationProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object"
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Forge.ReferenceCatalog.Tests | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
    "/api/catalog/items": {
      "post": {
        "operationId": "CreateCatalogItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatalogItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/catalog/items/{id}": {
      "get": {
        "operationId": "GetCatalogItem",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "tags": [
    {
      "name": "Catalog"
    }
  ]
}